+++ CHANGELOG.rst
@@ -21,6 +21,7 @@ x.x.x (x-x-x)
 * Migrated from `setup.py`` to `pyproject.toml` for packaging.
 * Added support for username/password authentication on external repositories during app creation.
 * Removed usage of `imp` module to better support python 3.12.
+* Added support for template services during app creation.
 
 3.17.2 (2023-03-16)
 -------------------